% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runitconv.R
\name{runitconv}
\alias{runitconv}
\title{Lauch the unit converter app}
\usage{
runitconv()
}
\value{
shiny application object
}
\description{
Lauch the unit converter app. A very simple and flexible app for unit
conversion based on the great \code{units} package.
}
