% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runitconv.R
\docType{package}
\name{runitconverter-package}
\alias{runitconverter}
\alias{runitconverter-package}
\title{RGPR: A package for processing and visualising ground-penetrating data
radar (GPR) data.}
\description{
The RGPR package provides two classes GPR and GPRsurvey
}
\author{
\strong{Maintainer}: Emanuel Huber \email{emanuel.huber@pm.me}

}
